+++ src/renderer/buffer/indexBuffer.h
@@ -11,7 +11,7 @@
 class IndexBuffer 
 {
 public:
-	IndexBuffer(const Device* device, const CommandBuffer* commandBuffers, const std::vector<uint16_t>& indices);
+	IndexBuffer(const Device* device, const CommandBuffer* commandBuffers, const std::vector<uint32_t>& indices);
 	~IndexBuffer();
 
 	inline VkBuffer GetIndexBuffer() const { return m_IndexBuffer; }
@@ -23,7 +23,7 @@ private:
 	const Device* m_Device;
 	const CommandBuffer* m_CommandBuffers; 
 
-	std::vector<uint16_t> m_Indices;
+	std::vector<uint32_t> m_Indices;
 
 	VkBuffer m_IndexBuffer;
 	VkDeviceMemory m_BufferMemory;
